Function to unify logger and ticker, updated gitignore
The new logger wraps the Google logger function and allows to define a message as well as the type (info, warn, error, and fatal). It outputs to both the log file (through logger) and to the system ticker by appending to an array, which discards the first element when a specified length is met. Cleaned up a few other elements as well. Comments pending.